pyproject.toml
[project]
name = "bus-check"
version = "0.1.0"
description = "Analyze CTA Frequent Network: ridership impact and headway adherence"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0",
    "requests>=2.31",
    "statsmodels>=0.14",
    "matplotlib>=3.8",
    "plotly>=5.18",
    "seaborn>=0.13",
    "jupyter>=1.0",
    "schedule>=1.2",
    "python-dotenv>=1.0",
    "csdid>=0.2.9",
    "drdid>=1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "responses>=0.25",
    "ruff>=0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bus_check"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
